"Add Redis to project for caching user shopping baskets"
This commit adds Redis to the project, primarily for the purpose of managing and caching user shopping baskets. A new 'BasketRepository' class has been implemented to interact with Redis. New entities 'CustomerBasket' and 'BasketItem' were created for managing baskets in Redis. Also added 'BasketController' to manage requests related to shopping baskets. These changes were necessary as the user's shopping basket was identified as a common piece of data that needs to be fetched and tweaked frequently, thus rendering it beneficial to being cached. The integration with Redis should speed up these operations.